Problem Description
Lexicographical Order, also known as alphabetic or dictionary order, orders characters as follows: 

For example, ball < cat, dog < dorm, Happy < happy, Zoo < ball.

A substring of a string is a contiguous block of characters in the string. For example, the substrings of abc are a, b, c, ab, bc, and abc.
Given a string, s , and an integer, k , complete the function so that it finds the lexicographically smallest and largest substrings of length .

Input Format

The first line contains a string denoting s. 
The second line contains an integer denoting k.

Constraints

 consists of English alphabetic letters only (i.e., [a-zA-Z]).
Output Format

Return the respective lexicographically smallest and largest substrings as a single newline-separated string.